+++ workspace/plugins/custom-fonts/src/components/Settings.svelte
@@ -63,6 +63,7 @@
 
     enum TabKey {
         base,
+        editor,
         code,
         graph,
         math,
@@ -105,6 +106,12 @@
                 text: i18n.settings.fontsSettings.base.title,
                 name: i18n.settings.fontsSettings.base.title,
                 icon: "🗛",
+            },
+            {
+                key: TabKey.editor,
+                text: i18n.settings.fontsSettings.editor.title,
+                name: i18n.settings.fontsSettings.editor.title,
+                icon: "🖺",
             },
             {
                 key: TabKey.code,
@@ -280,10 +287,58 @@
                 </Item>
             </div>
 
-            <!-- 标签页 2 - 代码字体设置 -->
+            <!-- 标签页 2 - 编辑器字体设置 -->
             <div
                 data-type={tabs.fonts[1].name}
                 class:fn__none={tabs.fonts[1].key !== focusTab}
+            >
+                <!-- 是否启用自定义字体列表 -->
+                <Item
+                    title={i18n.settings.fontsSettings.editor.enable.title}
+                    text={i18n.settings.fontsSettings.editor.enable.description}
+                >
+                    <Input
+                        slot="input"
+                        type={ItemType.checkbox}
+                        settingKey="enable"
+                        settingValue={config.fonts.editor.enable}
+                        on:changed={e => {
+                            config.fonts.editor.enable = e.detail.value;
+                            updated();
+                        }}
+                    />
+                </Item>
+
+                <!-- 自定义字体列表 -->
+                <Item
+                    block={true}
+                    title={i18n.settings.fontsSettings.editor.fontsList.title}
+                    text={i18n.settings.fontsSettings.editor.fontsList.description}
+                >
+                    <Input
+                        slot="input"
+                        block={true}
+                        type={ItemType.textarea}
+                        height={textareaHeight}
+                        settingKey="list"
+                        settingValue={config.fonts.editor.list.join("\n")}
+                        placeholder={i18n.settings.fontsSettings.editor.fontsList.placeholder}
+                        on:changed={e => {
+                            if (e.detail.value === "") {
+                                config.fonts.editor.list = [];
+                            } else {
+                                config.fonts.editor.list = e.detail.value.split("\n");
+                            }
+                            updated();
+                        }}
+                    />
+                </Item>
+            </div>
+
+            <!-- 标签页 3 - 代码字体设置 -->
+            <div
+                data-type={tabs.fonts[2].name}
+                class:fn__none={tabs.fonts[2].key !== focusTab}
             >
                 <!-- 是否启用自定义字体列表 -->
                 <Item
@@ -328,10 +383,10 @@
                 </Item>
             </div>
 
-            <!-- 标签页 3 - 图表字体设置 -->
+            <!-- 标签页 4 - 图表字体设置 -->
             <div
-                data-type={tabs.fonts[2].name}
-                class:fn__none={tabs.fonts[2].key !== focusTab}
+                data-type={tabs.fonts[3].name}
+                class:fn__none={tabs.fonts[3].key !== focusTab}
             >
                 <!-- 是否启用自定义字体列表 -->
                 <Item
@@ -376,10 +431,10 @@
                 </Item>
             </div>
 
-            <!-- 标签页 4 - 数学字体设置 -->
+            <!-- 标签页 5 - 数学字体设置 -->
             <div
-                data-type={tabs.fonts[3].name}
-                class:fn__none={tabs.fonts[3].key !== focusTab}
+                data-type={tabs.fonts[4].name}
+                class:fn__none={tabs.fonts[4].key !== focusTab}
             >
                 <!-- 是否启用自定义字体列表 -->
                 <Item
@@ -424,10 +479,10 @@
                 </Item>
             </div>
 
-            <!-- 标签页 5 - 表情符号字体设置 -->
+            <!-- 标签页 6 - 表情符号字体设置 -->
             <div
-                data-type={tabs.fonts[4].name}
-                class:fn__none={tabs.fonts[4].key !== focusTab}
+                data-type={tabs.fonts[5].name}
+                class:fn__none={tabs.fonts[5].key !== focusTab}
             >
                 <!-- 是否启用自定义字体列表 -->
                 <Item