+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # 更改日志 | CHANGE LOG
 
+## 2023-10-27
+
+- 新增自定义编辑器字体列表功能 | Add custom editor font list function.
+
 ## 2023-10-26
 
 - 添加打开 Jupyter 页面功能 | Add the function of opening Jupyter page.